requirements.txt
aniso8601==9.0.1
attrs==21.4.0
Automat==20.2.0
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
firebase-admin==5.2.0
Flask==2.1.1
Flask-HTTPAuth==4.6.0
Flask-RESTful==0.3.9
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.45.0
grpcio-status==1.45.0
gunicorn==20.1.0
h11==0.13.0
httpcore==0.12.3
httplib2==0.20.4
httpx==0.17.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.1
json-rpc==1.13.0
jsonrpc==3.0.1
MarkupSafe==2.1.1
msgpack==1.0.3
mypy-extensions==0.4.3
pjrpc==1.5.0
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
Twisted==22.4.0
typeguard==2.13.3
typing-inspect==0.7.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zope.interface==5.4.0
